+++ plugin/src/main/resources/index.jelly
@@ -1,4 +1,4 @@
 <?jelly escape-by-default='true'?>
 <div>
-    to push data to remote microservice
+    push data to remote microservice
 </div>
