--- plugin/src/main/resources/io/jenkins/plugins/ext/DemoPublisher/config.jelly
@@ -1,6 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <f:description>
-        to push data to remote microservice
-    </f:description>
-</j:jelly>
